// src/app/interfaces/parameter-maintenance-sys.ts
import { IdDescription } from './id-description';

export class ParameterMainSysInputHelp {
  PrmUnitShortName: string;
  PrmUnitAdd1: string;
  PrmUnitAdd2: string;
  PrmUnitAdd3: string;
  PrmUnitPhone: string;
  ExpenseGLDesc: string;
  PrmSystemPath: string;
  PrmBackUpPath: string;
  PrmDataPath: string;
  PrmEmailDataPath: string;
  PrmOutDataPath: string;
  PrmInDataPath: string;
  PrmPicturePath: string;
  PrmTimeOut: number;
}